Change type of pos to size_t

+++ src/hotspot/share/runtime/arguments.cpp
@@ -1206,7 +1206,7 @@ bool Arguments::process_settings_file(const char* file_name, bool should_exist, 
   }
 
   char token[1024];
-  int  pos = 0;
+  size_t pos = 0;
 
   bool in_white_space = true;
   bool in_comment     = false;
@@ -1215,7 +1215,7 @@ bool Arguments::process_settings_file(const char* file_name, bool should_exist, 
   bool result         = true;
 
   int c_or_eof = getc(stream);
-  while (c_or_eof != EOF && pos < (int)(sizeof(token) - 1)) {
+  while (c_or_eof != EOF && pos < (sizeof(token) - 1)) {
     // We have checked the c_or_eof for EOF. getc should only ever return the
     // EOF or an unsigned char converted to an int. We cast down to a char to
     // avoid the char to int promotions we would otherwise do in the comparisons
